requirements.txt
# Core agent stack (aligned constraints)
langgraph>=0.2.20,<0.3
langchain-core>=0.3,<0.4
langchain-anthropic>=0.3,<0.4

# Clients and drivers
anthropic>=0.34
appium-python-client>=5.2
selenium>=4.25

# OCR
pytesseract>=0.3.13
Pillow>=10.4
